Align repos

+++ build.sh
@@ -1,10 +1,20 @@
 #!/bin/bash
 #
 # Define how to build the libraries and executables:
-BUILD_TYPE=Debug
+if [[ -z "${BUILD_TYPE}" ]]; then
+  BUILD_TYPE="Debug"
+fi
+echo "Using BUILD_TYPE = $BUILD_TYPE"
+
+if [[ -z "${ENABLE_PFUNIT}" ]]; then
+  ENABLE_PFUNIT=ON
+fi
+echo "Using ENABLE_PFUNIT = $ENABLE_PFUNIT"
+
 Fortran_COMPILER=gfortran
 LIBSUFFIX="so"
 GENERATOR_FLAGS=""
+
 if [[ "$OSTYPE" == "linux-gnu" ]]; then
   LIBSUFFIX="so"
 elif [[ "$OSTYPE" == "darwin"* ]]; then
@@ -23,24 +33,28 @@ git submodule update --init --recursive
 #                                Build pFUnit                                  #
 #                                                                              #
 ################################################################################
-cd pFUnit || exit || exit
-# Create the build directory if it does not exist
-if [[ ! -d "build" ]]; then
-  mkdir build
+if [[ "$ENABLE_PFUNIT" == "ON" ]]; then
+  cd pFUnit || exit
+  # Create the build directory if it does not exist
+  if [[ ! -d "build" ]]; then
+    mkdir build
+  else
+    rm -rf build/*
+  fi
+  cd build || exit
+  echo "Updating cmake"
+  export PFUNIT_DIR=..//pFUnit/build/installed
+  export FC=$Fortran_COMPILER
+  cmake -DSKIP_MPI=yes "$GENERATOR_FLAGS" ../
+  echo "Building pFUnit"
+  cmake --build .
+  cmake --install .
+  cd ../  || exit
+  echo "Leaving pFUnit"
+  cd ../  || exit
 else
-  rm -rf build/*
+  echo "ENABLE_PFUNIT is 'OFF'. Skipping. Set to 'ON' to add it."
 fi
-cd build || exit || exit
-echo "Updating cmake"
-export PFUNIT_DIR=..//pFUnit/build/installed
-export FC=$Fortran_COMPILER
-cmake -DSKIP_MPI=yes "$GENERATOR_FLAGS" ../
-echo "Building pFUnit"
-cmake --build .
-cmake --install .
-cd ../  || exit || exit
-echo "Leaving pFUnit"
-cd ../  || exit || exit
 ################################################################################
 #                                                                              #
 #                                Build libslam                                 #
@@ -52,10 +66,10 @@ if [[ ! -d "build" ]]; then
 else
   rm -rf build/*
 fi
-cd build || exit || exit
+cd build || exit
 echo "Executing cmake"
 
-cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DCMAKE_Fortran_COMPILER=$Fortran_COMPILER -DENABLE_OpenMP_SUPPORT=ON -DENABLE_POSTGRESQL_SUPPORT=ON -DENABLE_PFUNIT=ON "$GENERATOR_FLAGS" ../
+cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DCMAKE_Fortran_COMPILER=$Fortran_COMPILER -DENABLE_OpenMP_SUPPORT=ON -DENABLE_POSTGRESQL_SUPPORT=ON -DENABLE_PFUNIT="$ENABLE_PFUNIT" "$GENERATOR_FLAGS" ../
 echo "Building libslam"
 cmake --build .
 cmake --install .
@@ -64,8 +78,8 @@ if [[ $? -ne 0 ]]; then
     exit $?
 fi
 echo "Manually preparing 'lib' and 'include' directories"
-cd ../  || exit || exit
-ln -sf build/include include  || exit || exit
-ln -sf build/lib lib  || exit || exit
+cd ../  || exit
+ln -sf build/include include  || exit
+ln -sf build/lib lib  || exit
 echo "Leaving libslam"
-echo "Done"
+echo "Done with $BUILD_TYPE build"
